services/video-processing-service/process_video.py
import os
import subprocess
import time
import requests
from google.cloud import storage

# --- Configuration ---
# These would typically be loaded from environment variables.
GCS_BUCKET_NAME = os.environ.get("GCS_BUCKET_NAME", "edu-platform-video-storage-bucket")
CONTENT_SERVICE_URL = os.environ.get("CONTENT_SERVICE_URL", "http://content-service:3001/api/v1")
DOWNLOAD_PATH = "/tmp/downloads"
TRANSCODE_PATH = "/tmp/transcoded"

# Initialize GCS client
# In a real GCP environment (like GKE with Workload Identity),
# authentication is handled automatically.
storage_client = storage.Client()

def handle_video_request(message_body):
    """
    The main handler for incoming video processing requests.
    """
    lesson_id = message_body.get("lessonId")
    text_content = message_body.get("textContent")

    if not lesson_id or not text_content:
        print("Error: Malformed message received.")
        return

    print(f"[{lesson_id}] Starting video processing for lesson.")

    try:
        # 1. Generate video with Google Veo (Placeholder)
        original_video_path = generate_video_with_veo(lesson_id, text_content)
        print(f"[{lesson_id}] Successfully generated original video at: {original_video_path}")

        # 2. Transcode the video (Placeholder)
        transcoded_files = transcode_video(lesson_id, original_video_path)
        print(f"[{lesson_id}] Successfully transcoded video into {len(transcoded_files)} resolutions.")

        # 3. Upload all files to GCS
        video_urls = {}
        all_files_to_upload = [(original_video_path, "original")] + [(p, r) for p, r in transcoded_files]
        for file_path, resolution in all_files_to_upload:
            gcs_url = upload_to_gcs(lesson_id, file_path, resolution)
            video_urls[resolution] = gcs_url
            print(f"[{lesson_id}] Uploaded {resolution} to {gcs_url}")

        # 4. Update the Content Service
        # We'll assume the main video URL is the original one for now.
        main_video_url = video_urls.get("original")
        update_content_service(lesson_id, main_video_url)
        print(f"[{lesson_id}] Successfully updated content service for lesson.")

    except Exception as e:
        print(f"[{lesson_id}] An error occurred during video processing: {e}")
        # Here you would add logic to handle failures, e.g., send to a dead-letter queue.
    finally:
        # Clean up local files
        print(f"[{lesson_id}] Cleaning up local files.")
        # Add cleanup logic here
        pass


def generate_video_with_veo(lesson_id, text):
    """Placeholder for Google Veo API call."""
    print(f"[{lesson_id}] Calling Google Veo API with text: '{text[:30]}...'")
    time.sleep(5) # Simulate API call latency

    # Create a dummy video file to simulate the download
    os.makedirs(DOWNLOAD_PATH, exist_ok=True)
    file_path = os.path.join(DOWNLOAD_PATH, f"{lesson_id}_original.mp4")
    with open(file_path, "w") as f:
        f.write(f"This is a dummy video for lesson {lesson_id}")
    return file_path


def transcode_video(lesson_id, original_path):
    """Placeholder for transcoding video with FFmpeg."""
    os.makedirs(TRANSCODE_PATH, exist_ok=True)
    resolutions = {"720p": "1280x720", "480p": "854x480"}
    output_paths = []

    for name, res in resolutions.items():
        output_path = os.path.join(TRANSCODE_PATH, f"{lesson_id}_{name}.mp4")

        # This is how you would call ffmpeg in a real application
        # command = [
        #     "ffmpeg", "-i", original_path,
        #     "-vf", f"scale={res}",
        #     "-preset", "fast",
        #     output_path
        # ]
        # print(f"[{lesson_id}] Running FFmpeg command: {' '.join(command)}")
        # subprocess.run(command, check=True)

        # Simulate by creating a dummy file
        with open(output_path, "w") as f:
            f.write(f"Dummy transcoded video at {name}")
        output_paths.append((output_path, name))

    return output_paths


def upload_to_gcs(lesson_id, local_path, resolution):
    """Uploads a file to Google Cloud Storage."""
    bucket = storage_client.bucket(GCS_BUCKET_NAME)
    file_name = os.path.basename(local_path)
    blob_name = f"videos/{lesson_id}/{file_name}"

    blob = bucket.blob(blob_name)

    # In a real app, you'd upload the file from local_path
    # blob.upload_from_filename(local_path)

    # For the sandbox, we'll upload a string
    blob.upload_from_string(f"Simulated upload of {file_name}", content_type="video/mp4")

    return blob.public_url


def update_content_service(lesson_id, video_url):
    """Updates the lesson in the Content Service with the new video URL."""
    url = f"{CONTENT_SERVICE_URL}/lessons/{lesson_id}/video" # Assuming this endpoint exists
    payload = {"video_url": video_url}

    print(f"[{lesson_id}] Updating content service at {url} with payload: {payload}")
    # response = requests.patch(url, json=payload)
    # response.raise_for_status() # Raise an exception for non-2xx status codes
    # print(f"[{lesson_id}] Content service updated successfully with status: {response.status_code}")

    # Simulate success
    return True
