services/video-processing-service/requirements.txt
# Python client for RabbitMQ
pika==1.3.2

# Google Cloud Storage client library
google-cloud-storage==2.10.0

# For making HTTP requests to other services (e.g., Content Service)
requests==2.31.0
